% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_colors}
\alias{get_model_colors}
\title{Need a function to fix colors to specific model_names
Note the we could have multiple ensembles, and if this is the case, we need to provide
specific colors for those multiple Ensembles}
\usage{
get_model_colors(m_df, ensemble_colors = NULL)
}
\description{
Need a function to fix colors to specific model_names
Note the we could have multiple ensembles, and if this is the case, we need to provide
specific colors for those multiple Ensembles
}
