% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparaison_plot.R
\name{get_scen_comp_data}
\alias{get_scen_comp_data}
\title{Function to get the scenario_comparison data}
\usage{
get_scen_comp_data(
  rnd_num,
  model_data,
  method = c("sum_incidence", "overall_cum", "zero_cum"),
  ...
)
}
\description{
Function to get the scenario_comparison data
}
